[MOD-278] Updated Public Modules for armv6/7 Issue with Xcode 4.2
Apple changes ARCHS_STANDARD_32_BIT to only compile armv7, but our modules need to work with our Titanium Mobile apps that support armv6. We need to explicitly state what architectures should be supported, and when. I also went ahead and bumped the SDKROOTs to the generic "iphoneos" from the specific versions that were in use so that Xcode can use the latest version of the SDK that it has. Enjoy!